+++ R/fp_corr_td.R
@@ -1,16 +1,17 @@
-#' Frequency and phase correction of MRS data in the time-domain.
+#' Time-domain spectral registration.
 #' 
 #' An implementation of the method published by Near et al MRM 73:44-50 (2015).
 #' 
 #' @param mrs_data MRS data to be corrected
 #' @param ref optional MRS data to use as a reference, the first dynamic of 
 #' mrs_data is used if this argument is not supplied.
-#' @param xlim optional frequency range to perform optimisation.
+#' @param xlim optional frequency range to perform optimisation, set to NULL
+#' to use the full range.
 #' @param max_t truncate the FID when longer than max_t to reduce time taken
 #' @return a list containing the corrected data; phase and shift values in units
 #' of degrees and Hz respectively.
 #' @export
-fp_corr_td <- function(mrs_data, ref = NULL, xlim = NULL, max_t = 0.2) {
+tdsr <- function(mrs_data, ref = NULL, xlim = c(4, 0.5), max_t = 0.2) {
   # align to first dynamic if ref is not given
   if (is.null(ref)) ref <- get_dyns(mrs_data, 1)