+++ R/mol_parameters.R
@@ -12,7 +12,9 @@ get_mol_paras <- function(name, ...) {
 #' @export
 get_mol_names <- function() {
   funs <- ls(getNamespace("spant"), all.names = TRUE)
-  funs <- funs[!funs %in% c("get_mol_paras")]
+  funs <- funs[!funs %in% c("get_mol_paras", "get_acq_paras",
+                            "get_1h_brain_basis_paras")]
+  
   sub("_paras", "", sub("get_", "", funs[grep("get_.*_paras", funs)]))
 }
 
@@ -42,6 +44,7 @@ get_uncoupled_mol <- function(name, chem_shift, nucleus, scale_factor, lw, lg) {
 #' @export
 print.mol_parameters <- function(x, ...) {
   cat(paste(c("Name        : ", x$name, "\n")), sep = "")
+  cat(paste(c("Source      : ", x$source, "\n")), sep = "")
   cat(paste(c("Spin groups : ", length(x$spin_groups), "\n")), sep = "")
   for (n in 1:length(x$spin_groups)) {
     cat("\n")
@@ -55,14 +58,22 @@ print.mol_parameters <- function(x, ...) {
       cat("j-coupling matrix\n")
       rownames(x$spin_groups[[n]]$j_coupling_mat) <- x$spin_groups[[n]]$chem_shift
       colnames(x$spin_groups[[n]]$j_coupling_mat) <- x$spin_groups[[n]]$chem_shift
-      print(x$spin_groups[[n]]$j_coupling_mat)
+      j_mat <- x$spin_groups[[n]]$j_coupling_ma
+      j_mat[upper.tri(j_mat, diag = TRUE)] <- NA
+      j_mat[j_mat == 0] <- NA
+      #j_mat <- t(j_mat)t
+      print(j_mat, na.print = "-")
     }
   }
 }
 
 get_m_cr_ch2_paras <- function(lw = NULL, lg = 0) {
   if (is.null(lw)) lw = 2
-  get_uncoupled_mol("-CrCH2", 3.913, "1H", -2, lw, lg)
+  paras <- get_uncoupled_mol("-CrCH2", 3.913, "1H", -2, lw, lg)
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  paras$source = source
+  paras
 }
 
 get_ala_paras <- function(lw = NULL, lg = 0) {
@@ -80,7 +91,11 @@ get_ala_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Ala")
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13: 129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Ala", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -97,7 +112,11 @@ get_asp_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Asp")
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Asp", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -107,6 +126,11 @@ get_cr_paras <- function(lw = NULL, lg = 0) {
   paras <- get_uncoupled_mol("Cr", 3.027, "1H", 3, lw, lg)
   paras_b <- get_uncoupled_mol("Cr", 3.913, "1H", 2, lw, lg)
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras$source = source
   paras
 }
 
@@ -130,7 +154,41 @@ get_gaba_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "GABA")
+  
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13: 129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "GABA", source = source)
+  class(paras) <- "mol_parameters"
+  paras
+}
+
+get_gaba_jn_paras <- function(lw = NULL, lg = 0) {
+  if (is.null(lw)) lw = 2
+  nucleus <- rep("1H", 6)
+  chem_shift <- c(2.2840, 2.2840, 1.8880, 1.8880, 3.0130, 3.0130)
+  j_coupling_mat <- matrix(0, 6, 6)
+  j_coupling_mat[2,1] <- -15.938
+  j_coupling_mat[3,1] <- 7.678
+  j_coupling_mat[4,1] <- 6.980
+  j_coupling_mat[3,2] <- 6.980
+  j_coupling_mat[4,2] <- 7.678
+  j_coupling_mat[4,3] <- -15.000
+  j_coupling_mat[5,3] <- 8.510
+  j_coupling_mat[6,3] <- 6.503
+  j_coupling_mat[5,4] <- 6.503
+  j_coupling_mat[6,4] <- 8.510
+  j_coupling_mat[6,5] <- -14.062
+  
+  spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
+                       j_coupling_mat = j_coupling_mat, scale_factor = 1,
+                       lw = lw, lg = lg)
+  
+  source <- "J-Difference Editing of Gamma-Aminobutyric Acid (GABA): 
+              Simulated and Experimental Multiplet Patterns. MRM 2013; 
+              70:1183-1191."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "GABA", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -152,7 +210,11 @@ get_gln_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Gln")
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Gln", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -191,8 +253,11 @@ get_gsh_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
   paras <- list(spin_groups = list(spin_group_a, spin_group_b, spin_group_c),
-                name = "GSH")
+                name = "GSH", source = source)
   
   class(paras) <- "mol_parameters"
   paras
@@ -200,7 +265,13 @@ get_gsh_paras <- function(lw = NULL, lg = 0) {
 
 get_gly_paras <- function(lw = NULL, lg = 0) {
   if (is.null(lw)) lw = 2
-  get_uncoupled_mol("Gly", 3.548, "1H", 2, lw, lg)
+  paras <- get_uncoupled_mol("Gly", 3.548, "1H", 2, lw, lg)
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras$source <- source
+  paras
 }
 
 get_ins_paras <- function(lw = NULL, lg = 0) {
@@ -219,21 +290,30 @@ get_ins_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  paras <- list(spin_groups = list(spin_group_a), name = "Ins")
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Ins", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
 
 get_lip09_paras <- function(ft) {
-  get_uncoupled_mol("Lip09", 0.89, "1H", 3, 0.14 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("Lip09", 0.89, "1H", 3, 0.14 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_lip13a_paras <- function(ft) {
-  get_uncoupled_mol("Lip13a", 1.28, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("Lip13a", 1.28, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_lip13b_paras <- function(ft) {
-  get_uncoupled_mol("Lip13b", 1.28, "1H", 2, 0.089 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("Lip13b", 1.28, "1H", 2, 0.089 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_lip20_paras <- function(ft) {
@@ -242,23 +322,32 @@ get_lip20_paras <- function(ft) {
   paras_c <- get_uncoupled_mol("Lip20", 2.8, "1H", 0.87, 0.2 * ft / 1e6, 1)
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
   paras$spin_groups[[3]] <- paras_c$spin_groups[[1]]
+  paras$source <- "LCModel manual."
   paras
 }
 
 get_mm09_paras <- function(ft) {
-  get_uncoupled_mol("MM09", 0.91, "1H", 3, 0.14 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("MM09", 0.91, "1H", 3, 0.14 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_mm12_paras <- function(ft) {
-  get_uncoupled_mol("MM12", 1.21, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("MM12", 1.21, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_mm14_paras <- function(ft) {
-  get_uncoupled_mol("MM14", 1.43, "1H", 2, 0.17 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("MM14", 1.43, "1H", 2, 0.17 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_mm17_paras <- function(ft) {
-  get_uncoupled_mol("MM17", 1.67, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras <- get_uncoupled_mol("MM17", 1.67, "1H", 2, 0.15 * ft / 1e6, 1)
+  paras$source <- "LCModel manual."
+  paras
 }
 
 get_mm20_paras <- function(ft) {
@@ -269,6 +358,7 @@ get_mm20_paras <- function(ft) {
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
   paras$spin_groups[[3]] <- paras_c$spin_groups[[1]]
   paras$spin_groups[[4]] <- paras_d$spin_groups[[1]]
+  paras$source <- "LCModel manual."
   paras
 }
 
@@ -291,14 +381,24 @@ get_naa_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  paras <- list(spin_groups = list(spin_group_a,spin_group_b), name = "NAA")
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a,spin_group_b), name = "NAA",
+                source = source)
   class(paras) <- "mol_parameters"
   paras
 }
 
 get_naag_paras <- function(lw = NULL, lg = 0) {
   if (is.null(lw)) lw = 0.5
-  get_uncoupled_mol("NAAG", 2.042, "1H", 3, lw, lg)
+  paras <- get_uncoupled_mol("NAAG", 2.042, "1H", 3, lw, lg)
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153. Note, only the acetyl moiety is simualted."
+  
+  paras$source <- source
+  paras
 }
 
 get_pch_paras <- function(lw = NULL, lg = 0) {
@@ -328,7 +428,12 @@ get_pch_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "PCh")
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a, spin_group_b), name = "PCh",
+                source = source)
+  
   class(paras) <- "mol_parameters"
   paras
 }
@@ -338,12 +443,23 @@ get_pcr_paras <- function(lw = NULL, lg = 0) {
   paras <- get_uncoupled_mol("PCr", 3.029, "1H", 3, lw, lg)
   paras_b <- get_uncoupled_mol("PCr", 3.930, "1H", 2, lw, lg)
   paras$spin_groups[[2]] <- paras_b$spin_groups[[1]]
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras$source <- source
   paras
 }
 
 get_sins_paras <- function(lw = NULL, lg = 0) {
   if (is.null(lw)) lw = 2
-  get_uncoupled_mol("sIns", 3.34, "1H", 6, lw, lg)
+  paras <- get_uncoupled_mol("sIns", 3.34, "1H", 6, lw, lg)
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras$source <- source
+  paras
 }
 
 get_tau_paras <- function(lw = NULL, lg = 0) {
@@ -361,7 +477,11 @@ get_tau_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Tau")
+  
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Tau", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -379,7 +499,10 @@ get_lac_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
-  paras <- list(spin_groups = list(spin_group_a), name = "Lac")
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Lac", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -401,7 +524,11 @@ get_glu_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Glu")
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Glu", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -422,7 +549,11 @@ get_glc_paras <- function(lw = NULL, lg = 0) {
   spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
-  paras <- list(spin_groups = list(spin_group_a), name = "Glc")
+  
+  source <- "Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Glc", source = source)
   class(paras) <- "mol_parameters"
   paras
 }
@@ -468,8 +599,59 @@ get_gpc_paras <- function(lw = NULL, lg = 0) {
                        j_coupling_mat = j_coupling_mat, scale_factor = 1,
                        lw = lw, lg = lg)
   
+  source <- "Corrigendum: Proton NMR chemical shifts and coupling constants for brain metabolites.
+              NMR Biomed. 2000; 13:129-153."
+  
   paras <- list(spin_groups = list(spin_group_a, spin_group_b, spin_group_c),
-                name = "GPC")
+                name = "GPC", source = source)
+  class(paras) <- "mol_parameters"
+  paras
+}
+
+get_10spin_paras <- function(lw = NULL, lg = 0) {
+  if (is.null(lw)) lw = 2
+  nucleus <- rep("1H", 10)
+  chem_shift <- c(5.216, 3.519, 3.698, 3.395, 3.822, 3.826, 3.749, 1, 2, 3)
+  j_coupling_mat <- matrix(0, 10, 10)
+  j_coupling_mat[2,1] <- 3.8
+  j_coupling_mat[3,2] <- 9.6
+  j_coupling_mat[4,3] <- 9.4
+  j_coupling_mat[5,4] <- 9.9
+  j_coupling_mat[6,5] <- 1.5
+  j_coupling_mat[7,5] <- 6
+  j_coupling_mat[7,6] <- -12.1
+  
+  spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
+                       j_coupling_mat = j_coupling_mat, scale_factor = 1,
+                       lw = lw, lg = lg)
+  
+  source <- "made up molecule"
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Glc", source = source)
+  class(paras) <- "mol_parameters"
+  paras
+}
+
+get_9spin_paras <- function(lw = NULL, lg = 0) {
+  if (is.null(lw)) lw = 2
+  nucleus <- rep("1H", 9)
+  chem_shift <- c(5.216, 3.519, 3.698, 3.395, 3.822, 3.826, 3.749, 1, 2)
+  j_coupling_mat <- matrix(0, 9, 9)
+  j_coupling_mat[2,1] <- 3.8
+  j_coupling_mat[3,2] <- 9.6
+  j_coupling_mat[4,3] <- 9.4
+  j_coupling_mat[5,4] <- 9.9
+  j_coupling_mat[6,5] <- 1.5
+  j_coupling_mat[7,5] <- 6
+  j_coupling_mat[7,6] <- -12.1
+  
+  spin_group_a <- list(nucleus = nucleus, chem_shift = chem_shift, 
+                       j_coupling_mat = j_coupling_mat, scale_factor = 1,
+                       lw = lw, lg = lg)
+  
+  source <- "made up molecule"
+  
+  paras <- list(spin_groups = list(spin_group_a), name = "Glc", source = source)
   class(paras) <- "mol_parameters"
   paras
 }